requirements.txt
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
docker-pycreds==0.4.0
filelock==3.12.0
gitdb==4.0.10
GitPython==3.1.31
hydra-core==1.3.2
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
omegaconf==2.3.0
packaging==23.1
pathtools==0.1.2
protobuf==4.22.4
psutil==5.9.5
pyparsing==3.0.9
PyYAML==6.0
requests==2.30.0
scikit-learn==1.2.2
scipy==1.10.1
sentry-sdk==1.22.1
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
sympy==1.11.1
threadpoolctl==3.1.0
torch==2.0.0
torch-geometric==2.3.1
tqdm==4.65.0
typing-extensions==4.5.0
urllib3==1.26.15
wandb==0.15.2
